refaktorerer for mer lesbar kode pluss shouldcomponent update på EnkelUtbetaling for færre re-renders
